refactor: Move functions out of `livechat.js` (#30706)
