+++ apps/meteor/app/livechat/server/methods/saveInfo.ts
@@ -8,6 +8,7 @@ import { callbacks } from '../../../../lib/callbacks';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { methodDeprecationLogger } from '../../../lib/server/lib/deprecationWarningLogger';
 import { Livechat } from '../lib/Livechat';
+import { Livechat as LivechatTyped } from '../lib/LivechatTyped';
 
 declare module '@rocket.chat/ui-contexts' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
@@ -77,7 +78,9 @@ Meteor.methods<ServerMethods>({
 			delete guestData.phone;
 		}
 
-		await Promise.allSettled([Livechat.saveGuest(guestData), Livechat.saveRoomInfo(roomData)]);
+		// Note: type is right, but `check` converts it to something else
+		// Since the endpoint is going to be deprecated, we can live with this for a while
+		await Promise.allSettled([LivechatTyped.saveGuest(guestData as any, userId), Livechat.saveRoomInfo(roomData)]);
 
 		const user = await Users.findOne({ _id: userId }, { projection: { _id: 1, username: 1 } });
